Update README.md
This code defines a function fibonacci that generates a Fibonacci sequence up to a specified number n. The Fibonacci sequence is a series of numbers in which each number is the sum of the two preceding ones, typically starting with 0 and 1.

The fibonacci function takes the input n as the length of the sequence to be generated. It initializes the Fibonacci sequence with the first two numbers [0, 1] and then iteratively calculates the subsequent numbers by adding the last two numbers in the sequence. It continues this process until the length of the sequence reaches n. Finally, it returns the generated Fibonacci sequence up to n.

In the main block (if name == "main":), the script prompts the user to input the number of Fibonacci numbers to generate, converts the input to an integer, and then calls the fibonacci function with this input. It then prints the generated Fibonacci sequence.